requirements.txt
openai
python-docx
pandas
openpyxl
python-dotenv
Pillow
faiss-cpu
numpy
